Move test classes to ambry-cloud.
